requirements.txt
#
# This file is autogenerated by pip-compile with python 3.10
# To update, run:
#
#    pip-compile
#
asgiref==3.5.0
    # via django
django==4.0.3
    # via -r requirements.in
django-environ==0.8.1
    # via -r requirements.in
django-treenode==0.18.1
    # via -r requirements.in
lxml==4.8.0
    # via pikepdf
packaging==21.3
    # via
    #   pikepdf
    #   pytesseract
pikepdf==5.1.1
    # via -r requirements.in
pillow==9.0.1
    # via
    #   pikepdf
    #   pytesseract
pyparsing==3.0.7
    # via packaging
pytesseract==0.3.9
    # via -r requirements.in
six==1.16.0
    # via django-treenode
sqlparse==0.4.2
    # via django
